requirements.txt
click ==8.1.3
# torch ==1.10.2
pytorch-lightning ==1.6.3
torchmetrics ==0.8.2
neptune-client ==0.16.2
albumentations ==1.1.0
segmentation-models-pytorch
timm
scikit-learn ==1.1.1
scipy ==1.8.1
hydra-core
